refactor: Attendance Request Form into Multiple Sub Component Sections
