+++ components/attendance/attendanceRequest/styles.ts
@@ -1,0 +1,125 @@
+import { StyleSheet } from "react-native";
+
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  header: {
+    padding: 16,
+    flexDirection: "row",
+    alignItems: "center",
+    borderBottomWidth: 1,
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: "600",
+    marginLeft: 8,
+  },
+  backButton: {
+    padding: 4,
+  },
+  form: {
+    padding: 16,
+    flex: 1,
+  },
+  fieldContainer: {
+    marginBottom: 24,
+    position: "relative",
+  },
+  label: {
+    fontSize: 16,
+    fontWeight: "500",
+    marginBottom: 12,
+  },
+  reasonHeaderContainer: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginBottom: 4,
+  },
+  reasonLoader: {
+    marginLeft: 10,
+  },
+  refreshButton: {
+    marginLeft: 10,
+    padding: 4,
+  },
+  input: {
+    borderWidth: 1,
+    borderRadius: 16,
+    padding: 16,
+    fontSize: 14,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  inputText: {
+    fontSize: 14,
+    flex: 1,
+  },
+  calendarIcon: {
+    marginLeft: 8,
+  },
+  textarea: {
+    height: 100,
+    textAlignVertical: "top",
+  },
+  dropdownContainer: {
+    borderWidth: 1,
+    borderRadius: 16,
+    padding: 16,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignItems: "center",
+  },
+  checkboxContainer: {
+    padding: 4,
+  },
+  checkboxRow: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginBottom: 16,
+    gap: 8,
+  },
+  checkboxLabel: {
+    fontSize: 14,
+  },
+  sectionTitle: {
+    fontSize: 16,
+    fontWeight: "600",
+    marginBottom: 16,
+  },
+  dropdown: {
+    position: "absolute",
+    top: "100%",
+    left: 0,
+    right: 0,
+    borderWidth: 1,
+    borderRadius: 16,
+    zIndex: 1000,
+    marginTop: 4,
+    overflow: "hidden",
+  },
+  dropdownItem: {
+    padding: 16,
+  },
+  emptyDropdownItem: {
+    padding: 16,
+    alignItems: "center",
+  },
+  dropdownItemText: {
+    fontSize: 14,
+  },
+  saveButtonContainer: {
+    padding: 12,
+  },
+  saveButton: {
+    padding: 16,
+    borderRadius: 16,
+    alignItems: "center",
+  },
+  saveButtonText: {
+    fontSize: 16,
+    fontWeight: "500",
+  },
+});
